Fix spec + Merge entity and behavior back again
